Re-factored locations of files in TopSecondDegreeByCount. Grouped similar files under same directory
