+++ graphjet-core/src/main/java/com/twitter/graphjet/algorithms/RecommendationResponse.java
@@ -17,6 +17,8 @@
 
 package com.twitter.graphjet.algorithms;
 
+import com.twitter.graphjet.algorithms.counting.recommendationInfo.RecommendationInfo;
+
 /**
  * This class encapsulates the required response from a {@link RecommendationAlgorithm}.
  */
