+++ graphjet-core/src/main/java/com/twitter/graphjet/algorithms/counting/recommendationInfo/RecommendationInfo.java
@@ -15,7 +15,10 @@
  */
 
 
-package com.twitter.graphjet.algorithms;
+package com.twitter.graphjet.algorithms.counting.recommendationInfo;
+
+import com.twitter.graphjet.algorithms.RecommendationAlgorithm;
+import com.twitter.graphjet.algorithms.RecommendationType;
 
 /**
  * This interface specifies the required information from each of the recommendations returned by
