+++ graphjet-core/src/main/java/com/twitter/graphjet/algorithms/counting/recommendationInfo/RecommendationInfoTweet.java
@@ -14,25 +14,29 @@
  * limitations under the License.
  */
 
-package com.twitter.graphjet.algorithms;
+
+package com.twitter.graphjet.algorithms.counting.recommendationInfo;
 
 import java.util.Map;
 
 import com.google.common.base.Objects;
+
+import com.twitter.graphjet.algorithms.RecommendationType;
 import it.unimi.dsi.fastutil.longs.LongList;
 
-/**
- * This class specifies user recommendations based on user follow, mention, and mediatag
- */
-public class RecommendationInfoUser implements RecommendationInfo {
+public class RecommendationInfoTweet implements RecommendationInfo {
   private final long recommendation;
   private final RecommendationType recommendationType;
   private final double weight;
   private final Map<Byte, LongList> socialProof;
 
-  public RecommendationInfoUser(long recommendation, double weight, Map<Byte, LongList> socialProof) {
+  /**
+   * This class specifies the tweet recommendation.
+   */
+  public RecommendationInfoTweet(long recommendation, double weight,
+                                 Map<Byte, LongList> socialProof) {
     this.recommendation = recommendation;
-    this.recommendationType = RecommendationType.USER;
+    this.recommendationType = RecommendationType.TWEET;
     this.weight = weight;
     this.socialProof = socialProof;
   }
@@ -72,7 +76,8 @@ public class RecommendationInfoUser implements RecommendationInfo {
 
     RecommendationInfoTweet other = (RecommendationInfoTweet) obj;
 
-    return Objects.equal(getRecommendation(), other.getRecommendation())
+    return
+      Objects.equal(getRecommendation(), other.getRecommendation())
         && Objects.equal(getRecommendationType(), other.getRecommendationType())
         && Objects.equal(getWeight(), other.getWeight())
         && Objects.equal(getSocialProof(), other.getSocialProof());
@@ -81,10 +86,10 @@ public class RecommendationInfoUser implements RecommendationInfo {
   @Override
   public String toString() {
     return Objects.toStringHelper(this)
-        .add("recommendation", recommendation)
-        .add("recommendationType", recommendationType)
-        .add("weight", weight)
-        .add("socialProof", socialProof)
-        .toString();
+      .add("recommendation", recommendation)
+      .add("recommendationType", recommendationType)
+      .add("weight", weight)
+      .add("socialProof", socialProof)
+      .toString();
   }
 }
