+++ graphjet-core/src/main/java/com/twitter/graphjet/algorithms/counting/response/TopSecondDegreeByCountResponse.java
@@ -15,9 +15,9 @@
  */
 
 
-package com.twitter.graphjet.algorithms.counting;
+package com.twitter.graphjet.algorithms.counting.response;
 
-import com.twitter.graphjet.algorithms.RecommendationInfo;
+import com.twitter.graphjet.algorithms.counting.recommendationInfo.RecommendationInfo;
 import com.twitter.graphjet.algorithms.RecommendationResponse;
 import com.twitter.graphjet.algorithms.RecommendationStats;
