+++ graphjet-core/src/test/java/com/twitter/graphjet/algorithms/salsa/SalsaBitmaskTest.java
@@ -27,14 +27,14 @@ import org.junit.Test;
 
 import static org.junit.Assert.assertArrayEquals;
 
-import com.twitter.graphjet.algorithms.RecommendationInfo;
+import com.twitter.graphjet.algorithms.counting.recommendationInfo.RecommendationInfo;
 import com.twitter.graphjet.algorithms.RequestedSetFilter;
 import com.twitter.graphjet.algorithms.ResultFilter;
 import com.twitter.graphjet.algorithms.ResultFilterChain;
 import com.twitter.graphjet.algorithms.StaticBipartiteGraph;
 import com.twitter.graphjet.algorithms.TweetCardFilter;
 import com.twitter.graphjet.algorithms.TweetIDMask;
-import com.twitter.graphjet.algorithms.RecommendationInfoTweet;
+import com.twitter.graphjet.algorithms.counting.recommendationInfo.RecommendationInfoTweet;
 import com.twitter.graphjet.algorithms.salsa.fullgraph.Salsa;
 import com.twitter.graphjet.bipartite.api.BipartiteGraph;
 import com.twitter.graphjet.stats.NullStatsReceiver;
